Basic Functions

Basic Functions

LightGraphs.jl defines the following basic types and functionality:

LightGraphs.AbstractEdge

LightGraphs.AbstractEdgeIter

LightGraphs.AbstractGraph

Base.reverse

LightGraphs.add_edge!

LightGraphs.add_vertex!

LightGraphs.add_vertices!

LightGraphs.all_neighbors

LightGraphs.common_neighbors

LightGraphs.degree

LightGraphs.degree_histogram

LightGraphs.density

LightGraphs.dst

LightGraphs.edges

LightGraphs.edgetype

LightGraphs.has_edge

LightGraphs.has_self_loops

LightGraphs.has_vertex

LightGraphs.in_neighbors

LightGraphs.indegree

LightGraphs.is_directed

LightGraphs.is_ordered

LightGraphs.isgraphical

LightGraphs.ne

LightGraphs.neighborhood

LightGraphs.neighbors

LightGraphs.num_self_loops

LightGraphs.nv

LightGraphs.out_neighbors

LightGraphs.outdegree

LightGraphs.rem_edge!

LightGraphs.rem_vertex!

LightGraphs.squash

LightGraphs.src

LightGraphs.vertices

LightGraphs.weights

LightGraphs.Δ

LightGraphs.Δin

LightGraphs.Δout

LightGraphs.δ

LightGraphs.δin

LightGraphs.δout

LightGraphs.AbstractEdge — Type.

AbstractEdge

An abstract type representing a single edge between two vertices of a graph.

source

LightGraphs.AbstractEdgeIter — Type.

AbstractEdgeIter

An abstract type representing an edge iterator.

source

LightGraphs.AbstractGraph — Type.

AbstractGraph

An abstract type representing a graph.

source

Base.reverse — Method.

reverse(e)

Create a new edge from e with source and destination vertices reversed.

source

LightGraphs.add_edge! — Method.

add_edge!(g, e)

Add a new edge e to g. Return false if add fails (e.g., if vertices are not in the graph, or edge already exists), true otherwise.

source

LightGraphs.add_vertex! — Method.

add_vertex!(g)

Add a new vertex to the graph g. Return true if the vertex was added successfully, false otherwise.

source

LightGraphs.dst — Method.

dst(e)

Return the destination vertex of edge e.

source

LightGraphs.edges — Method.

edges(g)

Return (an iterator to or collection of) the edges of a graph. For AbstractSimpleGraphs it returns a SimpleEdgeIter.

Implementation Notes

A returned iterator is valid for one pass over the edges, and is invalidated by changes to g.

source

LightGraphs.edgetype — Method.

edgetype(g)

Return the type of graph g's edge

source

LightGraphs.has_edge — Method.

has_edge(g, e)

Return true if the graph g has an edge e.

source

LightGraphs.has_vertex — Method.

has_vertex(g, v)

Return true if v is a vertex of g.

source

LightGraphs.in_neighbors — Method.

in_neighbors(g, v)

Return a list of all neighbors connected to vertex v by an incoming edge.

Implementation Notes

Returns a reference, not a copy. Do not modify result.

source

LightGraphs.is_directed — Method.

is_directed(g)

Return true if the graph is a directed graph; false otherwise.

source

LightGraphs.ne — Method.

ne(g)

Return the number of edges in g.

source

LightGraphs.nv — Method.

nv(g)

Return the number of vertices in g.

source

LightGraphs.out_neighbors — Method.

out_neighbors(g, v)

Return a list of all neighbors connected to vertex v by an outgoing edge.

Implementation Notes

Returns a reference, not a copy. Do not modify result.

source

LightGraphs.rem_edge! — Method.

rem_edge!(g, e)

Remove the edge e from g. Return false if edge removal fails (e.g., if edge does not exist), true otherwise.

source

LightGraphs.rem_vertex! — Method.

rem_vertex!(g, v)

Remove the vertex v from graph g. Return false if removal fails (e.g., if vertex is not in the graph), true otherwise.

source

LightGraphs.src — Method.

src(e)

Return the source vertex of edge e.

source

LightGraphs.vertices — Method.

vertices(g)

Return (an iterator to or collection of) the vertices of a graph.

Implementation Notes

A returned iterator is valid for one pass over the edges, and is invalidated by changes to g.

source

LightGraphs.add_vertices! — Method.

add_vertices!(g, n)

Add n new vertices to the graph g. Return true if all vertices were added successfully, false otherwise.

source

LightGraphs.all_neighbors — Function.

all_neighbors(g, v)

Return a list of all inbound and outbound neighbors of v in g. For undirected graphs, this is equivalent to both out_neighbors and in_neighbors.

Implementation Notes

Returns a reference, not a copy. Do not modify result.

source

LightGraphs.common_neighbors — Method.

common_neighbors(g, u, v)

Return the neighbors common to vertices u and v in g.

Implementation Notes

Returns a reference, not a copy. Do not modify result.

source

LightGraphs.degree — Function.

degree(g[, v])

Return a vector corresponding to the number of edges which start or end at each vertex in graph g. If v is specified, only return degrees for vertices in v. For directed graphs, this value equals the incoming plus outgoing edges. For undirected graphs, it equals the connected edges.

source

LightGraphs.degree_histogram — Function.

degree_histogram(g, degfn=degree)

Return a Dict with values representing the number of vertices that have degree represented by the key.

Degree function (for example, indegree or outdegree) may be specified by overriding degfn.

source

LightGraphs.density — Function.

density(g)

Return the density of g. Density is defined as the ratio of the number of actual edges to the number of possible edges ($|V|×(|V|-1)$ for directed graphs and $\frac{|V|×(|V|-1)}{2}$ for undirected graphs).

source

LightGraphs.has_self_loops — Method.

has_self_loops(g)

Return true if g has any self loops.

source

LightGraphs.indegree — Method.

indegree(g[, v])

Return a vector corresponding to the number of edges which end at each vertex in graph g. If v is specified, only return degrees for vertices in v.

source

LightGraphs.is_ordered — Method.

is_ordered(e)

Return true if the source vertex of edge e is less than or equal to the destination vertex.

source

LightGraphs.neighbors — Method.

neighbors(g, v)

Return a list of all neighbors reachable from vertex v in g. For directed graphs, the default is equivalent to out_neighbors; use all_neighbors to list inbound and outbound neighbors.

Implementation Notes

Returns a reference, not a copy. Do not modify result.

source

LightGraphs.num_self_loops — Method.

num_self_loops(g)

Return the number of self loops in g.

source

LightGraphs.outdegree — Method.

outdegree(g[, v])

Return a vector corresponding to the number of edges which start at each vertex in graph g. If v is specified, only return degrees for vertices in v.

source

LightGraphs.squash — Method.

squash(g)

Return a copy of a graph with the smallest practical type that can accommodate all vertices.

source

LightGraphs.weights — Method.

weights(g)

Return the weights of the edges of a graph g as a matrix. Defaults to DefaultDistance().

source

LightGraphs.Δ — Method.

Δ(g)

Return the maximum degree of vertices in g.

source

LightGraphs.Δin — Method.

Δin(g)

Return the maximum indegree of vertices in g.

source

LightGraphs.Δout — Method.

Δout(g)

Return the maximum outdegree of vertices in g.

source

LightGraphs.δ — Method.

δ(g)

Return the minimum degree of vertices in g.

source

LightGraphs.δin — Method.

δin(g)

Return the minimum indegree of vertices in g.

source

LightGraphs.δout — Method.

δout(g)

Return the minimum outdegree of vertices in g.

source

LightGraphs.isgraphical — Method.

isgraphical(degs)

Return true if the degree sequence degs is graphical, according to Erdös-Gallai condition.

Performance

Time complexity: ``\mathcal{O}(|degs|^2)``

source

LightGraphs.neighborhood — Method.

neighborhood(g, v, d)

Return a vector of the vertices in g at a geodesic distance less or equal to d from v.

Optional Arguments

dir=:out: If g is directed, this argument specifies the edge direction

with respect to v of the edges to be considered. Possible values: :in or :out.

source