Linear Algebra

Linear Algebra

LightGraphs.jl provides the following matrix operations on both directed and undirected graphs in the LinAlg submodule:

LightGraphs.adjacency_matrix — Function.

adjacency_matrix(g, dir=:out, T=Int)

Return a sparse adjacency matrix for a graph, indexed by [u, v] vertices. Non-zero values indicate an edge between u and v. Users may specify a direction (:in, :out, or :both are currently supported; :out is default for both directed and undirected graphs) and a data type for the matrix (defaults to Int).

Implementation Notes

This function is optimized for speed and directly manipulates CSC sparse matrix fields.

source

LightGraphs.adjacency_spectrum — Function.

Return the eigenvalues of the adjacency matrix for a graph g, indexed by vertex. Default values for dir and T are the same as those in adjacency_matrix.

Performance

Converts the matrix to dense with $nv^2$ memory usage.

Implementation Notes

Use eigs(adjacency_matrix(g); kwargs...) to compute some of the eigenvalues/eigenvectors.

source

LightGraphs.incidence_matrix — Function.

incidence_matrix(g, T=Int)

Return a sparse node-arc incidence matrix for a graph, indexed by [v, i], where i is in 1:ne(g), indexing an edge e. For directed graphs, a value of -1 indicates that src(e) == v, while a value of 1 indicates that dst(e) == v. Otherwise, the value is 0. For undirected graphs, both entries are 1 by default (this behavior can be overridden by the oriented optional argument).

Optional Arguments

oriented=false: If true, for an undirected graph g, the matrix will

contain arbitrary non-zero values representing connectivity between v and i.

source

LightGraphs.laplacian_matrix — Function.

laplacian_matrix(g, dir=:unspec, T=Int)

Return a sparse Laplacian matrix for a graph g, indexed by [u, v] vertices. For undirected graphs, dir defaults to :out; for directed graphs, dir defaults to :both. T defaults to Int for both graph types.

source

LightGraphs.laplacian_spectrum — Function.

laplacian_spectrum(g, dir=:unspec, T=Int)

Return the eigenvalues of the Laplacian matrix for a graph g, indexed by vertex. Default values for dir and T are the same as those in laplacian_matrix.

Performance

Converts the matrix to dense with $nv^2$ memory usage.

Implementation Notes

Use eigs(laplacian_matrix(g); kwargs...) to compute some of the eigenvalues/eigenvectors.

source

LightGraphs.spectral_distance — Function.

spectral_distance(G₁, G₂ [, k])

Compute the spectral distance between undirected n-vertex graphs G₁ and G₂ using the top k greatest eigenvalues. If k is ommitted, uses full spectrum.

References

JOVANOVIC, I.; STANIC, Z., 2014. Spectral Distances of Graphs Based on their Different Matrix Representations

source