Reading / Writing Graphs

Reading and writing Graphs

Graphs may be written to I/O streams and files using the savegraph function and read with the loadgraph function. The default graph format is a proprietary compressed LightGraphs.jl format lg. Other formats are available via the GraphPersistence package.

LightGraphs.loadgraph — Method.

loadgraph(file, gname="graph", format=LGFormat)

Read a graph named gname from file in the format format.

Implementation Notes

gname is graph-format dependent and is only used if the file contains multiple graphs; if the file format does not support multiple graphs, this value is ignored. The default value may change in the future.

source

LightGraphs.loadgraphs — Method.

loadgraphs(file, format=LGFormat())

Load multiple graphs from file in the format format. Return a dictionary mapping graph name to graph.

Implementation Notes

For unnamed graphs the default name "graph" will be used. This default may change in the future.

source

LightGraphs.savegraph — Method.

savegraph(file, g, gname="graph", format=LGFormat; compress=true)

Saves a graph g with name gname to file in the format format. If compress = true, use GZip compression when writing the file. Return the number of graphs written.

Implementation Notes

The default graph name assigned to gname may change in the future.

source

LightGraphs.savegraph — Method.

savegraph(file, g, d, format=LGFormat; compress=true)

Save a dictionary of graphname => graph to file in the format format. If compress = true, use GZip compression when writing the file. Return the number of graphs written.

Implementation Notes

Will only work if the file format supports multiple graph types.

source

Examples

save(STDOUT, g)
save("mygraph.jgz", g, "mygraph", compress=true)

savegraph("mygraph.jgz", g, compress=true)

dg = load("multiplegraphs.jgz") # dictionary of graphs