Making and Modifying Graphs

Making and Modifying Graphs

LightGraphs.jl provides a number of methods for creating a graph object, including tools for building and modifying graph objects, a wide array of graph generator functions, and the ability to read and write graphs from files (using GraphIO.jl).

Modifying graphs

LightGraphs.jl offers a range of tools for modifying graphs, including:

SimpleGraph
SimpleDiGraph
Edge
add_edge!
rem_edge!
add_vertex!
add_vertices!
rem_vertex!
zero

In addition to these core functions, more advanced operators can be found in Operators.

Graph Generators

LightGraphs.jl implements numerous graph generators, including random graph generators, constructors for classic graphs, numerous small graphs with familiar topologies, and random and static graphs embedded in Euclidean space.

LightGraphs.Edge

LightGraphs.StochasticBlockModel

LightGraphs.BinaryTree

LightGraphs.CliqueGraph

LightGraphs.CompleteBipartiteGraph

LightGraphs.CompleteDiGraph

LightGraphs.CompleteGraph

LightGraphs.CycleDiGraph

LightGraphs.CycleGraph

LightGraphs.DoubleBinaryTree

LightGraphs.Grid

LightGraphs.PathDiGraph

LightGraphs.PathGraph

LightGraphs.RoachGraph

LightGraphs.StarDiGraph

LightGraphs.StarGraph

LightGraphs.WheelDiGraph

LightGraphs.WheelGraph

LightGraphs.add_vertices!

LightGraphs.barabasi_albert

LightGraphs.barabasi_albert

LightGraphs.barabasi_albert!

LightGraphs.blockcounts

LightGraphs.erdos_renyi

LightGraphs.erdos_renyi

LightGraphs.euclidean_graph

LightGraphs.euclidean_graph

LightGraphs.expected_degree_graph

LightGraphs.kronecker

LightGraphs.make_edgestream

LightGraphs.random_configuration_model

LightGraphs.random_regular_digraph

LightGraphs.random_regular_graph

LightGraphs.random_tournament_digraph

LightGraphs.smallgraph

LightGraphs.static_fitness_model

LightGraphs.static_fitness_model

LightGraphs.static_scale_free

LightGraphs.static_scale_free

LightGraphs.stochastic_block_model

LightGraphs.stochastic_block_model

LightGraphs.watts_strogatz

Datasets

Other notorious graphs and integration with the MatrixDepot.jl package are available in the Datasets submodule of the companion package LightGraphsExtras.jl. Selected graphs from the Stanford Large Network Dataset Collection may be found in the SNAPDatasets.jl package.

All Generators

LightGraphs.StochasticBlockModel — Type.

StochasticBlockModel{T,P}

A type capturing the parameters of the SBM. Each vertex is assigned to a block and the probability of edge (i,j) depends only on the block labels of vertex i and vertex j.

The assignement is stored in nodemap and the block affinities a k by k matrix is stored in affinities.

affinities[k,l] is the probability of an edge between any vertex in block k and any vertex in block l.

Implementation Notes

Graphs are generated by taking random $i,j ∈ V$ and flipping a coin with probability affinities[nodemap[i],nodemap[j]].

source

LightGraphs.barabasi_albert! — Method.

barabasi_albert!(g::AbstractGraph, n::Integer, k::Integer)

Create a Barabási–Albert model random graph with n vertices. It is grown by adding new vertices to an initial graph g. Each new vertex is attached with k edges to k different vertices already present in the system by preferential attachment.

Optional Arguments

seed=-1: set the RNG seed.

source

LightGraphs.barabasi_albert — Method.

barabasi_albert(n::Integer, n0::Integer, k::Integer)

Create a Barabási–Albert model random graph with n vertices. It is grown by adding new vertices to an initial graph with n0 vertices. Each new vertex is attached with k edges to k different vertices already present in the system by preferential attachment. Initial graphs are undirected and consist of isolated vertices by default.

Optional Arguments

is_directed=false: if true, return a directed graph.

complete=false: if true, use a complete graph for the initial graph.

seed=-1: set the RNG seed.

source

LightGraphs.barabasi_albert — Method.

barabasi_albert(n, k)

Create a Barabási–Albert model random graph with n vertices. It is grown by adding new vertices to an initial graph with k vertices. Each new vertex is attached with k edges to k different vertices already present in the system by preferential attachment. Initial graphs are undirected and consist of isolated vertices by default.

Optional Arguments

is_directed=false: if true, return a directed graph.

complete=false: if true, use a complete graph for the initial graph.

seed=-1: set the RNG seed.

source

LightGraphs.blockcounts — Method.

blockcounts(sbm, A)

Count the number of edges that go between each block.

source

LightGraphs.erdos_renyi — Method.

erdos_renyi(n, ne)

Create an Erdős–Rényi random graph with n vertices and ne edges.

Optional Arguments

is_directed=false: if true, return a directed graph.

seed=-1: set the RNG seed.

source

LightGraphs.erdos_renyi — Method.

erdos_renyi(n, p)

Create an Erdős–Rényi random graph with n vertices. Edges are added between pairs of vertices with probability p.

Optional Arguments

is_directed=false: if true, return a directed graph.

seed=-1: set the RNG seed.

source

LightGraphs.expected_degree_graph — Method.

expected_degree_graph(ω)

Given a vector of expected degrees ω indexed by vertex, create a random undirected graph in which vertices i and j are connected with probability ω[i]*ω[j]/sum(ω).

Optional Arguments

seed=-1: set the RNG seed.

Implementation Notes

The algorithm should work well for maximum(ω) << sum(ω). As maximum(ω) approaches sum(ω), some deviations from the expected values are likely.

References

Connected Components in Random Graphs with Given Expected Degree Sequences, Linyuan Lu and Fan Chung. https://link.springer.com/article/10.1007%2FPL00012580

Efficient Generation of Networks with Given Expected Degrees, Joel C. Miller and Aric Hagberg. https://doi.org/10.1007/978-3-642-21286-4_10

source

LightGraphs.kronecker — Function.

kronecker(SCALE, edgefactor, A=0.57, B=0.19, C=0.19)

Generate a directed Kronecker graph with the default Graph500 parameters.

References

http://www.graph500.org/specifications#alg:generator

source

LightGraphs.make_edgestream — Method.

make_edgestream(sbm)

Take an infinite sample from the Stochastic Block Model sbm. Pass to Graph(nvg, neg, edgestream) to get a Graph object based on sbm.

source

LightGraphs.random_configuration_model — Method.

random_configuration_model(n, ks)

Create a random undirected graph according to the configuration model containing n vertices, with each node i having degree k[i].

Optional Arguments

seed=-1: set the RNG seed.

check_graphical=false: if true, ensure that k is a graphical sequence

(see isgraphical).

Performance

Time complexity is approximately $\mathcal{O}(n \bar{k}^2)$.

Implementation Notes

Allocates an array of $n \bar{k}$ Ints.

source

LightGraphs.random_regular_digraph — Method.

random_regular_digraph(n, k)

Create a random directed regular graph with n vertices, each with degree k.

Optional Arguments

dir=:out: the direction of the edges for degree parameter.

seed=-1: set the RNG seed.

Implementation Notes

Allocates an $n × n$ sparse matrix of boolean as an adjacency matrix and uses that to generate the directed graph.

source

LightGraphs.random_regular_graph — Method.

random_regular_graph(n, k)

Create a random undirected regular graph with n vertices, each with degree k.

Optional Arguments

seed=-1: set the RNG seed.

Performance

Time complexity is approximately $\mathcal{O}(nk^2)$.

Implementation Notes

Allocates an array of nk Ints, and . For $k > \frac{n}{2}$, generates a graph of degree $n-k-1$ and returns its complement.

source

LightGraphs.random_tournament_digraph — Method.

random_tournament_digraph(n)

Create a random directed tournament graph with n vertices.

Optional Arguments

seed=-1: set the RNG seed.

source

LightGraphs.static_fitness_model — Method.

static_fitness_model(m, fitness)

Generate a random graph with $|fitness|$ vertices and m edges, in which the probability of the existence of $Edge_{ij}$ is proportional to $fitness_i × fitness_j$.

Optional Arguments

seed=-1: set the RNG seed.

Performance

Time complexity is $\mathcal{O}(|V| + |E| log |E|)$.

References

Goh K-I, Kahng B, Kim D: Universal behaviour of load distribution in scale-free networks. Phys Rev Lett 87(27):278701, 2001.

source

LightGraphs.static_fitness_model — Method.

static_fitness_model(m, fitness_out, fitness_in)

Generate a random graph with $|fitness\_out + fitness\_in|$ vertices and m edges, in which the probability of the existence of $Edge_{ij}$ is proportional with respect to $i ∝ fitness\_out$ and $j ∝ fitness\_in$.

Optional Arguments

seed=-1: set the RNG seed.

Performance

Time complexity is $\mathcal{O}(|V| + |E| log |E|)$.

References

Goh K-I, Kahng B, Kim D: Universal behaviour of load distribution in scale-free networks. Phys Rev Lett 87(27):278701, 2001.

source

LightGraphs.static_scale_free — Method.

static_scale_free(n, m, α_out, α_in)

Generate a random graph with n vertices, m edges and expected power-law degree distribution with exponent α_out for outbound edges and α_in for inbound edges.

Optional Arguments

seed=-1: set the RNG seed.

finite_size_correction=true: determines whether to use the finite size correction

proposed by Cho et al.

Performance

Time complexity is $\mathcal{O}(|V| + |E| log |E|)$.

References

Goh K-I, Kahng B, Kim D: Universal behaviour of load distribution in scale-free networks. Phys Rev Lett 87(27):278701, 2001.

Chung F and Lu L: Connected components in a random graph with given degree sequences. Annals of Combinatorics 6, 125-145, 2002.

Cho YS, Kim JS, Park J, Kahng B, Kim D: Percolation transitions in scale-free networks under the Achlioptas process. Phys Rev Lett 103:135702, 2009.

source

LightGraphs.static_scale_free — Method.

static_scale_free(n, m, α)

Generate a random graph with n vertices, m edges and expected power-law degree distribution with exponent α.

Optional Arguments

seed=-1: set the RNG seed.

finite_size_correction=true: determines whether to use the finite size correction

proposed by Cho et al.

Performance

Time complexity is $\mathcal{O}(|V| + |E| log |E|)$.

References

Goh K-I, Kahng B, Kim D: Universal behaviour of load distribution in scale-free networks. Phys Rev Lett 87(27):278701, 2001.

Chung F and Lu L: Connected components in a random graph with given degree sequences. Annals of Combinatorics 6, 125-145, 2002.

Cho YS, Kim JS, Park J, Kahng B, Kim D: Percolation transitions in scale-free networks under the Achlioptas process. Phys Rev Lett 103:135702, 2009.

source

LightGraphs.stochastic_block_model — Method.

stochastic_block_model(c, n)

Return a Graph generated according to the Stochastic Block Model (SBM).

c[a,b] : Mean number of neighbors of a vertex in block a belonging to block b. Only the upper triangular part is considered, since the lower traingular is determined by $c[b,a] = c[a,b] * \frac{n[a]}{n[b]}$. n[a] : Number of vertices in block a

Optional Arguments

seed=-1: set the RNG seed.

For a dynamic version of the SBM see the StochasticBlockModel type and related functions.

source

LightGraphs.stochastic_block_model — Method.

stochastic_block_model(cint, cext, n)

Return a Graph generated according to the Stochastic Block Model (SBM), sampling from an SBM with $c_{a,a}=cint$, and $c_{a,b}=cext$.

source

LightGraphs.watts_strogatz — Method.

watts_strogatz(n, k, β)

Return a Watts-Strogatz small model random graph with n vertices, each with degree k. Edges are randomized per the model based on probability β.

Optional Arguments

is_directed=false: if true, return a directed graph.

seed=-1: set the RNG seed.

source

LightGraphs.BinaryTree — Method.

BinaryTree(k::Integer)

Create a binary tree of depth k.

source

LightGraphs.CliqueGraph — Method.

CliqueGraph(k, n)

Create a graph consisting of n connected k-cliques.

source

LightGraphs.CompleteBipartiteGraph — Method.

CompleteBipartiteGraph(n1, n2)

Create an undirected complete bipartite graph with n1 + n2 vertices.

source

LightGraphs.CompleteDiGraph — Method.

CompleteDiGraph(n)

Create a directed complete graph with n vertices.

source

LightGraphs.CompleteGraph — Method.

CompleteGraph(n)

Create an undirected complete graph with n vertices.

source

LightGraphs.CycleDiGraph — Method.

CycleDiGraph(n)

Create a directed cycle graph with n vertices.

source

LightGraphs.CycleGraph — Method.

CycleGraph(n)

Create an undirected cycle graph with n vertices.

source

LightGraphs.DoubleBinaryTree — Method.

BinaryTree(k::Integer)

Create a double complete binary tree with k levels.

References

Used as an example for spectral clustering by Guattery and Miller 1998.

source

LightGraphs.Grid — Method.

Grid(dims; periodic=false)

Create a $|dims|$-dimensional cubic lattice, with length dims[i] in dimension i.

Optional Arguments

periodic=false: If true, the resulting lattice will have periodic boundary

condition in each dimension.

source

LightGraphs.PathDiGraph — Method.

PathDiGraph(n)

Creates a directed path graph with n vertices.

source

LightGraphs.PathGraph — Method.

PathGraph(n)

Create an undirected path graph with n vertices.

source

LightGraphs.RoachGraph — Method.

RoachGraph(k)

Create a Roach Graph of size k.

References

Guattery and Miller 1998

source

LightGraphs.StarDiGraph — Method.

StarDiGraph(n)

Create a directed star graph with n vertices.

source

LightGraphs.StarGraph — Method.

StarGraph(n)

Create an undirected star graph with n vertices.

source

LightGraphs.WheelDiGraph — Method.

WheelDiGraph(n)

Create a directed wheel graph with n vertices.

source

LightGraphs.WheelGraph — Method.

WheelGraph(n)

Create an undirected wheel graph with n vertices.

source

LightGraphs.smallgraph — Method.

smallgraph(s)
smallgraph(s)

Create a small graph of type s. Admissible values for s are:

s	graph type

:bull	A bull graph.

:chvatal	A Chvátal graph.

:cubical	A Platonic cubical graph.

:desargues	A Desarguesgraph.

:diamond	A diamond graph.

:dodecahedral	A Platonic dodecahedral graph.

:frucht	A Frucht graph.

:heawood	A Heawood graph.

:house	A graph mimicing the classic outline of a house.

:housex	A house graph, with two edges crossing the bottom square.

:icosahedral	A Platonic icosahedral graph.

:krackhardtkite	A Krackhardt-Kite social network graph.

:moebiuskantor	A Möbius-Kantor graph.

:octahedral	A Platonic octahedral graph.

:pappus	A Pappus graph.

:petersen	A Petersen graph.

:sedgewickmaze	A simple maze graph used in Sedgewick's Algorithms in C++: Graph Algorithms (3rd ed.)

:tetrahedral	A Platonic tetrahedral graph.

:truncatedcube	A skeleton of the truncated cube graph.

:truncatedtetrahedron	A skeleton of the truncated tetrahedron graph.

:truncatedtetrahedron_dir	A skeleton of the truncated tetrahedron digraph.

:tutte	A Tutte graph.

source

LightGraphs.euclidean_graph — Method.

euclidean_graph(points)

Given the d×N matrix points build an Euclidean graph of N vertices and return a graph and Dict containing the distance on each edge.

Optional Arguments

L=1: used to bound the d dimensional box from which points are selected.

p=2

bc=:open

Implementation Notes

Defining the d-dimensional vectors x[i] = points[:,i], an edge between vertices i and j is inserted if norm(x[i]-x[j], p) < cutoff. In case of negative cutoff instead every edge is inserted. For p=2 we have the standard Euclidean distance. Set bc=:periodic to impose periodic boundary conditions in the box $[0,L]^d$.

source

LightGraphs.euclidean_graph — Method.

euclidean_graph(N, d; seed=-1, L=1., p=2., cutoff=-1., bc=:open)

Generate N uniformly distributed points in the box $[0,L]^{d}$ and return a Euclidean graph, a map containing the distance on each edge and a matrix with the points' positions.

source